Implementing UI functions and Documenting UI.
Moved UI Functions to separate UIControl.c/.h files to clean-up main and ease documentation.
WriteRecipe was changed to saveRecipe to avoid confusion with the eeprom function. This function was also fully implemented to take user inptus

nameSearch was modified to properly search to properly access and search the name.

dispenseSpice: Added validation checks for the requested and remaining quantity.

dispenseRecipe: Added validation checks for each spice against the requested vs remaining quantity.

RecipeList was modified to be a 2D array with pre-allocated memory for up to MAXNAMESIZE characters. This was also corrected to use the MAXNUMRECP (max number of recipes). Since we know the pre-defined size of the names and memory  utilization is not an issue, this implementation eases the need for having to manage memory.

printRecipe: This is a work in progress. There are issues with sprintf corrupting memory so some investigation is required