Added refillSpice and ChangeSpice command
WIP:
This commit pushes additional commands as well as updates some commands to provide clarity

The commands array and minfields array has been replaced with the commands structure (UICmdStructType) This allows for the min fields to be coupled to the command to ease maintainability.
The identifyCommand was updated to support this change.

refillSpice: Function adds capability to refill a specified spice with a given quantity.

changeSpice: Function adds capability to  change the name of a spice as well as update the quantity.

UIRackHome: Function simply calls the StepRackHome function and checks for any errors to indicate to user.

StepRackHome: Bug found that rotations was not actually completing  a full 360. Corrected by multiplying with gear ratio. Also made the HOME_OFFSET a global variable as wellas the SVO_ENG_POS and SVO_DIS_POS globals for mods during debug. This is in place in case we want to add support for changing of these values from a "Settings" page in the UI.